[refactor/#179] SignUp 개편 및 관련 코드 수정
(Squash -> Create and merge 방식 전환)

+++ module-auth/src/main/java/com/inhabas/api/auth/domain/oauth2/member/dto/UpdateRequestDto.java
@@ -11,7 +11,7 @@ import java.util.List;
 @AllArgsConstructor
 public class UpdateRequestDto {
 
-    private List<Integer> memberIdList;
+    private List<Long> memberIdList;
 
     @Schema(example = "pass, fail")
     private String state;
